 edge-ai-project/requirements.txt
# Core
torch>=1.13.0
torchvision>=0.14.0

# ONNX & runtime
onnx>=1.12.0
onnxruntime>=1.15.0
onnxruntime-tools>=1.6.0

# Quantization helpers
onnxruntime-extensions>=0.1.5

# Web demo
gradio>=3.16.0
Pillow>=9.0.0
numpy>=1.23.0

# Optional (for measurement)
psutil>=5.9.0
